% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.CP.county.R
\name{detect.CP.county}
\alias{detect.CP.county}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
detect.CP.county(
  dat = list(),
  state.show = NULL,
  county.show = NULL,
  plot.show = TRUE,
  cp.sd = 2,
  slope.r = 5,
  manual.decision = FALSE
)
}
\description{
This function detects the abnormal data point and/or data period of county level covid-19 data (or similar type of data including the count time series and epidemic data).
}
