% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.DR.county.R
\name{detect.DR.county}
\alias{detect.DR.county}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
detect.DR.county(
  dat = list(),
  state.show = NULL,
  county.show = NULL,
  log.test = FALSE,
  plot.show = FALSE,
  test = "wo"
)
}
\description{
This function detects the delay reporting issue of county level covid-19 data (or similar type of data including the count time series and epidemic data).
}
