% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.DR.R
\name{detect.DR}
\alias{detect.DR}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
detect.DR(
  dat = list(),
  level = "county",
  state.show = NULL,
  county.show = NULL,
  log = FALSE,
  plot.show = FALSE,
  test = "wo"
)
}
\arguments{
\item{dat}{A data frame or list containing the time series of infection or death cases.
\cr}

\item{level}{Level of data for detection. Level "county" represents county level data, and level "state" means state level data. Default is set to "county".
\cr}

\item{state.show}{Within the dataset, which state the detection algorithm should focus on. Default is set to \code{NULL}.
\cr}

\item{county.show}{For county level data, which county the detection algorithm should focus on. Default is set to \code{NULL}.
\cr}

\item{log}{Whether perform logarithm transform for the original data. Default is set to \code{FALSE}.
\cr}

\item{plot.show}{Whether output plots to the console for sequence that is tested to have delay-reporting issue. Default is set to \code{FALSE}.
\cr}

\item{test}{The test algorithm used for delay reporting detection. The available tests include the QS test (test = "qs"), Friedman test (test = "fried"), Kruskall-Wallis (test = "kw"), F-test on seasonal dummies (test = "seasdum") or the Welch test (test = "welch"). Default is set to be "wo".
\cr}
}
\value{
A list containing the following information.
\cr
\code{dat.sub} The data which contains the information on the state/county specified.
\cr
\code{dr.res} A dataframe with the last column indicating whether the state/county has delay-reporting issue (0=No, 1=Yes).
\cr
}
\description{
This function detects the delay reporting issue of county/state level covid-19 data (or similar type of data including the count time series and epidemic data).
}
\details{
This R package is the implementation program for article entitled "Comparing and Integrating US COVID-19 Data from Multiple Sources: A County-Level Dataset with Local Features" by Guannan Wang, Zhiling Gu, Xinyi Li, Shan Yu, Myungjin Kim, Yueying Wang, Lei Gao, and Li Wang.
}
\examples{

}
