% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.DR.state.R
\name{detect.DR.state}
\alias{detect.DR.state}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
detect.DR.state(
  dat = list(),
  state.show = NULL,
  log.test = FALSE,
  plot.show = FALSE,
  test = "wo"
)
}
\description{
This function detects the order-dependency violation of state level covid-19 data (or similar type of data including the count time series and epidemic data).
}
