% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.OD.county.R
\name{detect.OD.county}
\alias{detect.OD.county}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
detect.OD.county(
  dat = list(),
  state.show = NULL,
  county.show = NULL,
  plot.show = NULL
)
}
\description{
This function detects the order-dependency violation of county level covid-19 data (or similar type of data including the count time series and epidemic data).
}
