% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.OD.R
\name{detect.OD}
\alias{detect.OD}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
detect.OD(
  dat = list(),
  level = "county",
  state.show = NULL,
  county.show = NULL,
  plot.show = NULL
)
}
\arguments{
\item{dat}{A data frame or list containing the time series of infection or death cases.
\cr}

\item{level}{Level of data for detection. Level "county" represents county level data, and level "state" means state level data. Default is set to "county".
\cr}

\item{state.show}{Within the dataset, which state the detection algorithm should focus on.
\cr}

\item{county.show}{For county level data, which county the detection algorithm should focus on.
\cr}
}
\value{
A list containing the following information.
\cr
\code{dat.sub} The data which contains the information on the state/county specified.
\cr
\code{dat.new} The data after the order-dependency violation detection. \code{NA} represents the data point that needs repairment.
\cr
\code{n.OD} Number of order-dependency violation.
\cr
\code{dates.OD} The dates in the time series that the order-dependency violation.
\cr
\code{diff} The daily new increasing infection/death cases for COVID-19 in US.
\cr
}
\description{
This function detects the order-dependency violation of county level covid-19 data (or similar type of data including the count time series and epidemic data).
}
\details{
This R package is the implementation program for article entitled "Comparing and Integrating US COVID-19 Daily Data from Multiple Sources: A County-Level Dataset with Local Characteristics" by Guannan Wang, Zhiling Gu, Xinyi Li, Shan Yu, Myungjin Kim, Yueying Wang, Lei Gao, and Li Wang.
}
\examples{

}
