% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.OD.state.R
\name{detect.OD.state}
\alias{detect.OD.state}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
detect.OD.state(dat = list(), state.show = NULL, plot.show = NULL)
}
\description{
This function detects the order-dependency violation of state level covid-19 data (or similar type of data including the count time series and epidemic data).
}
