% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cdcar.R
\name{plot.cdcar}
\alias{plot.cdcar}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
\method{plot}{cdcar}(dat = list(), level = "state")
}
\arguments{
\item{dat}{A data frame or list containing the time series to plot.
\cr}

\item{level}{Level of data repairment. Level "county" represents county level data repairment, and level "state" means state level data repairment. Default is set to "county".
\cr}
}
\description{
This function plot the time series data of covid-19 data.
}
\details{
This R package is the implementation program for article entitled "Comparing and Integrating US COVID-19 Daily Data from Multiple Sources: A County-Level Dataset with Local Characteristics" by Guannan Wang, Zhiling Gu, Xinyi Li, Shan Yu, Myungjin Kim, Yueying Wang, Lei Gao, and Li Wang.
}
\examples{

}
