% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair.ARMA.county.R
\name{repair.ARMA.county}
\alias{repair.ARMA.county}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
repair.ARMA.county(dat.I = list(), dat.D = NULL, h = 7)
}
\description{
This function repairs the county level covid-19 data (or similar type of data including the count time series and epidemic data) using autoregressive moving average model of count time series.
}
