% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair.cdcar.R
\name{repair.cdcar}
\alias{repair.cdcar}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
repair.cdcar(
  dat.I = list(),
  dat.D = NULL,
  h = 7,
  level = "county",
  method = "AR"
)
}
\arguments{
\item{dat.I}{A data frame or list containing the time series of infection cases. \code{NA} in the dataset indicates the data points that need repairment.
\cr}

\item{dat.D}{A data frame or list containing the time series of death cases. Default sets to \code{NULL}. Repair method "STEM" requires both infection cases and death cases for the repairment.
\cr}

\item{h}{Parameter used in data repairment. For repair methods "AR" and "ARMA", this parameter controls the number of lags in time series modeling. For repair methods "CLEP" and "STEM", this parameter controls the time window for model fitting.
\cr}

\item{level}{Level of data repairment. Level "county" represents county level data repairment, and level "state" means state level data repairment. Default is set to "county".
\cr}

\item{method}{Method used for data repairment. "AR" represents autoregressive model of count time series, "ARMA" represents the autoregressive moving average models of count time series, "CLEP" represents the combined linear and exponential method, "STEM" represents the spatio-temporal epidemic method. Default is set to "AR".}
}
\value{
A data frame or list containing the repaired time series, including the infection cases and/or death cases.
}
\description{
This function repairs the abnormal data in covid-19 data (or similar type of data including the count time series and epidemic data).
}
\details{
This R package is the implementation program for article entitled "Comparing and Integrating US COVID-19 Data from Multiple Sources: A County-Level Dataset with Local Features" by Guannan Wang, Zhiling Gu, Xinyi Li, Shan Yu, Myungjin Kim, Yueying Wang, Lei Gao, and Li Wang.
}
\examples{

}
