% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair.CLEP.county.R
\name{repair.CLEP.county}
\alias{repair.CLEP.county}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
repair.CLEP.county(dat.I = list(), dat.D = NULL, h = 5)
}
\description{
This function repairs the county level covid-19 data (or similar type of data including the count time series and epidemic data) using CLEP method.
}
