% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair.JP.state.R
\name{repair.JP.state}
\alias{repair.JP.state}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
repair.JP.state(
  dat.I.JP = list(),
  dat.D.JP = list(),
  dat.I = list(),
  dat.D = list(),
  h = 7,
  date.start = as.Date("2020-03-01")
)
}
\description{
This function repairs the state level covid-19 data (or similar type of data including the count time series and epidemic data) using autoregressive model of count time series.
}
