% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair.STEM.state.R
\name{repair.STEM.state}
\alias{repair.STEM.state}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
repair.STEM.state(dat.I = list(), dat.D = NULL, h = 7)
}
\description{
This function repairs the state level covid-19 data (or similar type of data including the count time series and epidemic data) using STEM method.
}
