% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.DR.county.R
\name{score.DR.county}
\alias{score.DR.county}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
score.DR.county(
  dat = list(),
  state.show = NULL,
  county.show = NULL,
  log.test = FALSE,
  test = "wo"
)
}
\arguments{
\item{dat}{A data frame or list containing the time series of infection or death cases.
\cr}

\item{state.show}{Within the dataset, which state the detection algorithm should focus on. Default is set to \code{NULL}.
\cr}

\item{county.show}{For county level data, which county the detection algorithm should focus on. Default is set to \code{NULL}.
\cr}

\item{test}{The test algorithm used for delay reporting detection. The available tests include the QS test (test = "qs"), Friedman test (test = "fried"), Kruskall-Wallis (test = "kw"), F-test on seasonal dummies (test = "seasdum") or the Welch test (test = "welch"). Default is set to be "wo".
\cr}

\item{level}{Level of data for detection. Level "county" represents county level data, and level "state" means state level data. Default is set to "county".
\cr}

\item{log}{Whether perform logarithm transform for the original data. Default is set to \code{FALSE}.
\cr}
}
\value{
A list containing the following information.
\cr
\code{dat.sub} The data which contains the information on the state/county specified.
\cr
\code{score} A dataframe with the second column (r.DR) indicating the percentage of counties that are detected to have delay reporting issue for each state, and the third column (score) indicating the score based on the percentage with highest "A" and lowest "C".
\cr
}
\description{
This function detects the delay reporting issue of county/state level covid-19 data (or similar type of data including the count time series and epidemic data).
}
\details{
This R package is the implementation program for article entitled "Comparing and Integrating US COVID-19 Data from Multiple Sources: A County-Level Dataset with Local Features" by Guannan Wang, Zhiling Gu, Xinyi Li, Shan Yu, Myungjin Kim, Yueying Wang, Lei Gao, and Li Wang.
}
\examples{

}
