% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.OD.county.R
\name{score.OD.county}
\alias{score.OD.county}
\title{COVID-19 Data Checking and Reparing (CDCAR)}
\usage{
score.OD.county(dat = list(), state.show = NULL, county.show = NULL)
}
\arguments{
\item{dat}{A data frame or list containing the time series of infection or death cases.
\cr}

\item{state.show}{Within the dataset, which state the detection algorithm should focus on.
\cr}

\item{county.show}{For county level data, which county the detection algorithm should focus on.
\cr}
}
\value{
A list containing the following information.
\cr
\code{dat.sub} The data which contains the information on the state/county specified.
\cr
\code{score} The score for the dataset.
\cr
}
\description{
This function provides the score of a dataset based on the order-dependency violation of county level covid-19 data (or similar type of data including the count time series and epidemic data).
}
\details{
This R package is the implementation program for article entitled "Comparing and Integrating US COVID-19 Data from Multiple Sources: A County-Level Dataset with Local Features" by Guannan Wang, Zhiling Gu, Xinyi Li, Shan Yu, Myungjin Kim, Yueying Wang, Lei Gao, and Li Wang.
}
\examples{

}
